--- clients/apple/Shared/Config/Development.xcconfig
@@ -1,2 +1,0 @@
-// Local development
-API_LOCATION = http:/$()/localhost:8000
