--- clients/apple/Shared/Config/Production.xcconfig
@@ -1,2 +1,0 @@
-// API Location
-API_LOCATION = https:/$()/api.prod.lockbook.net
